user authorization with pundit working

+++ app/controllers/bookmarks_controller.rb
@@ -2,6 +2,7 @@ class BookmarksController < ApplicationController
   def index
     @topic = Topic.find(params[:topic_id])
   end
+
   def show
     @bookmark = Bookmark.find(params[:id])
   end
@@ -12,44 +13,46 @@ class BookmarksController < ApplicationController
   end
 
   def edit
-      @bookmark = Bookmark.find(params[:id])
+    @bookmark = Bookmark.find(params[:id])
   end
 
   def create
-  @bookmark = Bookmark.new
-  @bookmark.url = params[:bookmark][:url]
-   @topic = Topic.find(params[:topic_id])
-   @bookmark.topic = @topic
-  if @bookmark.save
-    flash[:notice] = "Bookmark was saved."
-    redirect_to [@topic, @bookmark]
-  else
-    flash.now[:alert] = "There was an error saving the bookmark. Please try again."
-    render :new
-  end
+    @bookmark = Bookmark.new
+    @bookmark.user = current_user
+    @bookmark.url = params[:bookmark][:url]
+    @topic = Topic.find(params[:topic_id])
+    @bookmark.topic = @topic
+    if @bookmark.save
+      flash[:notice] = 'Bookmark was saved.'
+      redirect_to [@topic, @bookmark]
+    else
+      flash.now[:alert] = 'There was an error saving the bookmark. Please try again.'
+      render :new
+    end
 end
 
-def update
+  def update
     @bookmark = Bookmark.find(params[:id])
+    authorize @bookmark
     @bookmark.url = params[:bookmark][:url]
 
     if @bookmark.save
-      flash[:notice] = "Bookmark was updated."
+      flash[:notice] = 'Bookmark was updated.'
       redirect_to [@bookmark.topic, @bookmark]
     else
-      flash.now[:alert] = "There was an error saving the bookmark. Please try again."
+      flash.now[:alert] = 'There was an error saving the bookmark. Please try again.'
       render :edit
     end
-  end
+    end
 
   def destroy
     @bookmark = Bookmark.find(params[:id])
-
+    authorize @bookmark
     if @bookmark.destroy
       flash[:notice] = "\"#{@bookmark.url}\" was deleted successfully."
-       redirect_to @bookmark.topic
+      redirect_to @bookmark.topic
     else
-      flash.now[:alert] = "There was an error deleting the bookmark."
+      flash.now[:alert] = 'There was an error deleting the bookmark.'
       render :show
     end
   end
